refactor: 프로그레스 바 공통 컴포넌트화
