--- frontend/apps/client/src/app/(main)/recommend/_components/index.ts
@@ -1,1 +1,0 @@
-export { default as FlowHeader } from './FlowHeader';
