+++ frontend/apps/client/src/shared/components/SegmentedProgress.tsx
