+++ frontend/apps/client/src/shared/components/index.ts
@@ -2,12 +2,14 @@ export { default as AudioPlayer } from './AudioPlayer';
 export { default as BackButton } from './BackButton';
 export { default as BaseModal } from './BaseModal';
 export { default as Calendar } from './Calendar';
+export { default as FlowHeader } from './FlowHeader';
 export { default as GenreCard } from './GenreCard';
 export { default as PageBanner } from './PageBanner';
 export { default as PageHeader } from './PageHeader';
 export { default as ReportCard } from './ReportCard';
 export { default as SectionHeader } from './SectionHeader';
 export { default as SectionTab } from './SectionTab';
+export { default as SegmentedProgress } from './SegmentedProgress';
 export { default as SelectChip } from './SelectChip';
 export { default as Sidebar } from './Sidebar';
 export { default as SituationCard } from './SituationCard';
